Mostly reverted support for 2 leb128 encodings in lfsr_data_t
As much as it was a nice way to utilize all 96-bits of lfsr_data_t, the
added code and RAM cost just made this not worth it.

The main problem with inlining leb128s into lfsr_data_t is that
lfsr_data_t reads have to contend with a number of awkward corner cases
involving offsets into the not-yet-encoded leb128s.

Adding to this the extra overhead encoding the number of leb128s in
lfsr_data_t, and the fact that 2 leb128s is mostly useless when metadata
redundancy > 2, I'm reverting this back to only a single optional leb128
in lfsr_data_ts limited to lfsr_bd_progdata:

   0 = in-device buffer     1 = on-disk data
  .----+----+----+----.  .----+----+----+----.
  |0|      size       |.>|1|      size       |
  |----+----+----+----|  |----+----+----+----|
  | (optional leb128) |  |        off        |
  |----+----+----+----|  |----+----+----+----|
  |       buffer      |  |       buffer      |
  '----+----+----+----'  '----+----+----+----'

It's also worth noting that since this leb128 is limited to
lfsr_bd_progdata, it shouldn't add any code cost to readonly variants of
littlefs.

Another interesting thing is that, while this single-injected-leb128-
for-lfsr_bd_progdata sounds limited on paper, it covers a number of
convenient use cases:

- injecting directory-ids into name attributes
- programming single leb128s, such as bookmarks
- prefixing B-tree branches with weights (not-yet implemented)