Made btree commit track bids correctly during recursion
This doesn't have that big an impact at the moment, but limiting the
bids/rids to well intentioned values helps development and debugging.

As we tail-recurse up the btree, the current bid always indicates the
left-most/least id in the current rbyd. This contrasts with pid, which
is the right-most id in the current rbyd. Before this the bid was
somewhat arbitrary after the first leaf, which risks confusion later.

This also implies bid=0 when we reach the root, which is a useful debug
assertion.