Some more btree cleanup
Trying to deduplicate the attr-list constructions before tail recursion
as much as possible.

Also tried rearranging the fetch of our parent to after we commit to our
current rbyd. This makes more logical sense, in terms of the order of
operations, but does mean duplicate parent fetches for the different
code paths...

            code          stack
  before:  20830           1752
  after:   20750 (-0.4%)   1752 (+0.0%)