More btree tweaks, now with better handling of degenerate parents
Normally, in btrees, the height of the btree only decreases when nodes
are merged.

But not in our btree! Thanks again to lazy merging, btree nodes can be
dropped instead of merged.

We don't have enough information to decrease the height of the btree
exactly when we drop a btree node, since we don't know how many siblings
the original node had, but we can at least decrease the height of the
btree if we notice this condition during normal commits.