Leveraged lfsr_rbyd_appendall's rid range to avoid needing the bid
The start_rid already adjusts the attrs relative rid, and we don't
have a reason to store -1 rid tags in our btree yet, so we can drop the
extra bid parameters.

This is nice because the bid parameter was leaking across abstraction
layers a bit. Our rbyds don't need to understand bids anywhere else.

Unfortunately, this did require some tweaks to btree split, which was
expecting to be able to temporarily write out a -1 rid tag in the case
we're creating a new root. But this is a good change, the -1 rid tag
was sort of a hack that risked other similar bugs.

Funnily enough, these two changes canceled each other out exactly:

            code          stack
  before:  20626           1712
  after:   20626 (+0.0%)   1712 (+0.0%)