Dropped LFSR_NO_REBALANCE for now
This was not updated after changing btree merges to compact, and as more
of this codebase depends on our rebalanced code size, it's not worth
keeping this code around right now.

Maybe the option for not rebalancing during compaction should be looked
into again in the future. But I'll leave that up to then.