Rearranged lfsr_mkdir to avoid tracking an imaginary mid
lfsr_mkdir creates two mid entries atomically, one for the bookmark and
one for the actual dir entry. It looks up where to insert both, which is
necessary for some other checks, and then inserts one mid while tracking
the position of the other mid using our opened-mdir subsystem.

But it's a bit of a challenge to track an mid that hasn't been created
yet.

Consider what happends if we create a bookmark immediately adjacent to
our dir entry:

  0.0 bookmark parent  ->  0.0 bookmark parent  ->  0.0 bookmark parent
  (tracking 0.1)           0.1 bookmark child       0.1 bookmark child
                           (tracking 0.2)           0.2 dir child

That's not right.

To fix this we could add some special handling to our opened-mdir
subsystem to track opened-mdirs specially if they don't actually exist.

Or, as it turns out, just create the dir/bookmark entries in the
opposite order. Which happens to avoid this problem completely:

  0.0 bookmark parent  ->  0.0 bookmark parent  ->  0.0 bookmark parent
  (tracking 0.1)           0.1 dir child            0.1 dir child
                           (tracking 0.2)           0.2 bookmark child

The reason this works is that, thanks to our bookmarks, we can never
actually have a bookmark immediately preceding our dir entry. The
imaginary dir entry will always be preceded by either our parent or
other entry thanks to ordering by did first:

  (tracking 0.0)       ->  (tracking 0.0)       ->  0.0 bookmark child
  0.0 bookmark parent      0.0 bookmark parent      0.1 bookmark parent
  (tracking 0.1)           0.1 dir child            0.2 dir child

Previously we relied on rid=-1 to handle this as a special case, but
this workaround would stop working with compressed mids, where rid=-1
becomes unrepresentable.